==> 03_r_stata_datamanagement.Rmd <==
---
output:
  html_document: default
  pdf_document: default
---

```{r include=TRUE, echo=FALSE}
knitr::opts_chunk$set(engine.path = list(
  stata = '/Applications/Stata/StataBE.app/Contents/MacOS/StataBE'
))
library(Statamarkdown)
```

# Data management

## Renaming variables
Variables names are not always explicit and you might want to rename some.

::::{style="display: grid; grid-template-columns: 1fr 1fr; grid-column-gap: 30px;"}

:::{}
### R{-}
In **R** *data.frame*, variables are in column and to retrieve their names you can use the *colnames()* function. It will return a vector which values can be accessed and modify using indexing.
```{r, echo=FALSE}
load("sleepApnea-clean.Rdata")
```

```{r, eval = FALSE}
colnames(sleep)
colnames(sleep)[1] <- "identifier"
```
Note : *rownames()* does also exist if you need to access and manipulate the individuals
:::
  
:::{}
### Stata{-}
In **Stata**, one way is to use the *rename* function.
```{stata, eval=F}
rename number identifier
```
:::
::::

## Missing values

Missing values are always problematic. First, because it is often difficult to know why they are missing: not answered, not attributable, lost in follow-up, miss recording...They can be largely missing for an individuals as well as for a specific variable. 



::::{style="display: grid; grid-template-columns: 1fr 1fr; grid-column-gap: 30px;"}

:::{}
### R{-}
In **R**, you can assess the number missing values with the *summary()* function or the *is.na()* conditional testing function.
```{r, eval = FALSE}
is.na(sleep$age)
```
A basic and crude method is to replace missing value by a value such as the average value. For the variable age, let's replace the missing value by the mean age of sample using the indexing technique.
```{r, eval = TRUE}
# create a fake missing value
# for individual #1
sleep$age[1] <- NA
sleep$age[1:5]
na_id <- is.na(sleep$age)
sleep$age[na_id] <- mean(sleep$age,
                         na.rm=TRUE)
round(sleep$age[1:5],0)
```
:::
  
:::{}
### Stata{-}
```{stata first-Stata, echo=FALSE, include=FALSE, collectcode=TRUE}
import delimited using sleepApnea-clean
```

In **Stata** an overview of the missing data can be done as follow:
```{stata}
misstable summarize
misstable patterns
```
To replace one missing value by the mean of the variable, do
```{stata}
egen meanage = mean(age)
replace age = meanage if missing(age)
```
:::
::::

Above a certain threshold that varies, from 4% to 20%, researchers try to do multiple imputation of missing values. Although there is no perfect solution, there are basically 3 scenario and associated imputation techniques. Data are missing completely at random (MCAR) or data are missing at random (MAR) or data are missing not at random (MNAR). For the  latest, it would be difficult to perform a legitimate analysis. This advanced topic is not covered in MPH1 class but see below for references if needed.

For multiple imputation with **R** see the package [*Mice*](https://cran.r-project.org/web/packages/mice/index.html) for more sophisticated missing values imputation

For multiple imputation with **Stata** see [Rose Meideiros ](https://www.stata.com/meeting/switzerland16/slides/medeiros-switzerland16.pdf) presentation.

## Create new variables

### Changing scale
In epidemiology, you may have to transform and change the scale of your variables. For instance, a logarithm transformation might help approximate a Normal distribution for some skewed continuous variable. You might also want to normalized or standardized (sometimes called z-score) your data. When you normalize, you   bring all the values between 0 and 1, while preserving the distances between the values. When you standardize, you bring all the values between 0 and 1 with a mean $\mu$ to 0 and a standard deviation $\sigma$ to 1.

::::{style="display: grid; grid-template-columns: 1fr 1fr; grid-column-gap: 10px;"}
:::{}
### R {-}
In **R** the transformation can only be some simple computation:

```{r, eval=FALSE}
sleep$age_log <- log(sleep$age)
sleep$age_sqr <- sleep$age^2
```

To normalize or standardize, you can do: 
```{r, eval=FALSE}
# normalization
minage <- sleep$age 
maxage <- sleep$age
sleep$age_norm<- 
  (sleep$age-minage)/(maxage-minage)
# standardization (z-score)
sleep$age_std<- 
  (sleep$age-mean(sleep$age))/sd(sleep$age)
```

_Note_: be careful if you have missing values you should add  *na.rm=TRUE* as argument in your function calls.
:::

:::{}
### Stata {-}
The syntax are really similar with **Stata**
```{stata, eval=FALSE}
generate age_log = ln(age)
generate age_sqr = age^2
```

```{stata, eval=FALSE}
// normalization
generate age_norm = (age-23)/74-23
// standardization
generate age_std = (age-47.69)/10.69
```
:::
::::


### From continuous to categorical
In epidemiology and data sciences more globally, you often need to convert numerical variable into to groups, i.e. a categorical variable. Here is an example with age as a continuous variable that you split into age groups.

::::{style="display: grid; grid-template-columns: 1fr 1fr; grid-column-gap: 10px;"}
:::{}
### R{-}
In **R**, you can use the function *cut()* to split into intervals (equal or not) a quantitative variable. The 3 main arguments of the function are:  

- breaks : interval using number as boundaries
- include.lowest: whether or not you include the lowest value if the data series
- right : whether or not you include the upper-bound value of the interval in your group (inclusion is symbolized by `]` and exclusion by `)`.
```{r}
sleep$age_group <- cut(sleep$age, 
                       breaks = c(20,40,60,80),
                       include.lowest=TRUE, 
                       right = FALSE)
table(sleep$age_group)
```
:::
  
:::{}
### Stata{-}
In **Stata** the syntax is quite similar
```{stata}
egen age_group = cut(age),at(20,40,60,80)
tabulate age_group
```
:::
::::
  

### Dummy variables

A dummy variable is one that takes only the value 0 or 1 to indicate the absence or presence of some categorical effect that may be expected to shift the outcome.

Dummy variable are often create as the result of a conditional expression. The table below presents the differents symbols avaible in **R** and **Stata** for conditional testing.
```{r, echo=FALSE}
symb <- data.frame("Sign"=c("==", "!=", ">", "<", ">=", "<=", "&", "|"),
                   "Definition"= c("equality", "not equality", "greater than", "less than", "greater than or equal to", "less than or equal to", "and (both statements are true)", "or (either statement is true)"),
                   "TRUE"=c("3 == 3", "3 != 4", "4 > 2", "3 < 4", "4 >= 3", "4 <= 5", "(4 > 2) & (3 == 3))", "(3 == 2) | (1 <= 2)"),
                   "FALSE"=c( "3 == 4", "3 != 3", "3 > 4", "4 < 2", "4 >= 5", "4 <= 3", "(4 > 2) & (1 > 2)", "(4 < 2) | (1 > 2)" ))

knitr::kable(symb, align = "c")
```

::::{style="display: grid; grid-template-columns: 1fr 1fr; grid-column-gap: 10px;"}
:::{}
 
### R{-}

The combination of any of the above symbols with the *ifelse()* function useful for creating a dummy variable based on  conditional testing as you can assign 0/1 values (or any specified values) to your variable depending on whether the element of test is TRUE or FALSE. As example, we create below 

```{r}
sleep$obese <- ifelse(sleep$bmi > 30,  1, 0)
```

Note that with **R** there is more than one way of doing things and if you adopt the use of  *tidyverse* suite of packages or the *data.table* library you might use other functions. Here are examples:
```{r,eval=FALSE}
# with dplyr
sleep <- sleep 
%>% mutate("obese"= as.numeric(bmi>25))
#with data.table
sleep <- data.table(sleep) 
sleep <- sleep[ , obese:=ifelse(bmi > 25,  1, 0)]
```

```{r, echo=FALSE}
save(sleep, file="sleepApnea-clean.Rdata")
write.table(sleep, file="sleepApnea-clean.csv", sep=";", row.names = FALSE)
```
:::
  
:::{}
### Stata{-}
In **Stata** the commands are similar: 
```{stata, eval=FALSE}
generate obese = 0 
replace obese = 1 if bmi>30
//handling possible missing values
replace obese = . if missing(bmi)
```
Or you can do
```{stata, eval=FALSE}
gen obese = bmi>30 if !missing(bmi)
```
:::
::::
  
::::{style="display: grid; grid-template-columns: 1fr 1fr; grid-column-gap: 10px;"}
:::{}
### R{-}

In **R** and the library tidyverse, you can transform a long variable to wide. You can return boolean data to create dummy variable or even use other values from another variable.  
```{r, eval=FALSE}
sleep <- sleep %>% mutate(i=1) 
      %>% spread(gender, i, fill=0)
```
:::

:::{}
### Stata{-}
In **Stata** the function tabulate with the generate() option will generate a whole sets of dummy variables.

If you type as below you will see a frequency table of how many times gender takes on each of those values. 
```{stata, eval=FALSE}
tabulate gender
```
If you type as below you will see the table, and tabulate will create variable names male, female that take on values 1 and 0, male being 1 when gender==male, female being 1 when gender==female.
```{stata, eval=FALSE}
tabulate group, generate(g)
```
:::
::::

## Merging data set

Merging or joining data set can be particularly useful when enriching your data from multiple sources.

In the example below, we merge the results of a questionnaire on daily concentration level for each patient ([here](DayConcentration.csv) to download) with the sleep apnea data .

::::{style="display: grid; grid-template-columns: 1fr 1fr; grid-column-gap: 10px;"}
:::{}
### R{-}

In **R** there is more than way way of doing things. The base function is *merge()*.
```{r}
dayconcentration <- read.table("DayConcentration.csv", header=TRUE, sep="\t")
# the commun key do not have the same label so it need to be specified
sleep <- merge(sleep, dayconcentration, by.x="number", by.y="id")
sleep[1:5, c("number","age", "concentration_mark")]  
```
You can use the all.x=TRUE or all.y=TRUE arguments to specify the merging mode. With tidyverse, you can also use different convenient functions *left_join()*, *right_join()*, and *full_join()*
```{r, eval=FALSE}
# all values of the sleep data set are kept 
# non match in dayconcentration are ignored
sleep <- left_join(sleep, dayconcentration, by=c ("number"="id"))
# all values of dayconcentration are kept 
sleep <-right_join(sleep, dayconcentration, by=c ("number"="id"))
# all values are kept 
sleep <-full_join(sleep, dayconcentration, by=c ("number"="id"))
```
Note that the common key can be composed of more than one variable. See help pages for details.
:::

:::{}
### Stata{-}
In **Stata** the merging is less trivial than in R. The two datasets cannot be opened in Stata at the same time and **the common (key) variables must have the same name**.

Make sure that:

- both datasets are sorted by all the ids and save as dta  
- one dataset is loaded into Stata (in this case sleep),  
- use the merge function but make sure to map where the using data is located (in this case for example “c:\folders\data\dayconcentration.dta”)

```{stata, eval=FALSE}
merge  number using dayconcentration
```

:::
::::
